对数据库结构进行了修改，对整体架构进行了一些调整

+++ src/main/java/com/mind_map/controller/UserController.java
@@ -64,7 +64,7 @@ public class UserController {
         // 得到用户的账号密码
         String username = user.getUsername();
         String password = user.getPassword();
-        Long id = user.getId();
+        Integer id = user.getId();
 
         // 查看账号是否已存在
         LambdaQueryWrapper<User> queryWrapper = new LambdaQueryWrapper<>();
